requirements.txt
Django
django-ckeditor
django-macros
django-extensions
Pillow
Markdown
python-dotenv
